requirements.txt

asn1crypto==0.23.0
Babel==2.5.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.7.4
botocore==1.10.4
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.7
configparser==3.5.0
coverage==4.4.1
cryptography==2.0.3
decorator==4.1.2
dill==0.2.7.1
dj-stripe==1.0.0.post1
Django==1.11.5
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.1
django-extra-views==0.6.4
django-fontawesome==0.3.1
django-haystack==2.6.1
django-livereload-server==0.2.3
django-localflavor==1.5.2
django-material==1.1.1
django-oscar==1.5.1
django-phonenumber-field==1.3.0
django-polymorphic==1.3
django-redis==4.8.0
django-secure==1.0.1
django-storages==1.6.5
django-tables2==1.11.0
django-treebeard==4.1.2
django-webtest==1.9.2
django-widget-tweaks==1.4.1
dnspython==1.15.0
docutils==0.14
email-validator==1.0.2
execnet==1.4.1
factory-boy==2.9.2
Faker==0.8.3
flake8==3.4.1
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-auth==1.1.1
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.15
google-resumable-media==0.2.3
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.6.3
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
isort==4.2.15
jmespath==0.9.3
jsonfield==2.0.2
mccabe==0.6.1
mock==2.0.0
monotonic==1.3
oauth2client==3.0.0
olefile==0.44
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
phonenumberslite==8.8.1
pickleshare==0.7.4
Pillow==4.2.1
pluggy==0.5.2
ply==3.8
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
psycopg2==2.7.3.1
ptyprocess==0.5.2
purl==1.3.1
py==1.4.34
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycodestyle==2.3.1
pycountry==17.5.14
pycparser==2.18
pyflakes==1.5.0
Pygments==2.2.0
pyOpenSSL==17.3.0
pyprof2calltree==1.4.3
pysolr==3.6.0
pytest==3.2.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-xdist==1.20.0
python-dateutil==2.6.1
python-doc-inherit==0.3.0
python-dotenv==0.7.1
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.13
scandir==1.5
selenium==3.11.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.4a1
sqlparse==0.2.3
stripe==1.53.0
tenacity==4.4.0
tornado==4.5.2
tox==2.8.2
tqdm==4.15.0
traitlets==4.3.2
transifex-client==0.12.4
Unidecode==0.4.21
urllib3==1.22
uWSGI==2.0.15
virtualenv==15.1.0
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.3
WebTest==2.0.28
Werkzeug==0.12.2
whitenoise==3.3.0
Whoosh==2.7.4
